Augment error message when there are non-matching columns. (#468)
This makes it easier to understand why an asset has an error. In real life, an asset had non-ascii characters in the column name, so add a specific check for that, because when the character is non-printable, it's impossible to tell with the naked eye why the name is wrong.