Allow zero-score tracts to be displayed in the table (#473)
The table was using zero score to indicate that a block group was filtered out. That's normally true (we set the score to 0 for any filtered-out block group), but if the filtering options are set to include everything, some block groups that have a genuine zero score won't be included.

Also sneak in a fix to score_bounds_map_test since Google Maps changed the "Zoom in" button to not be visible by default.

Fixes #472